Implementation of ServerProperties and the Parsing in IrcClient.
It exposes the Server Features which are sent by the Server with Reply Code 005 (RPL_ISUPPORT / RPL_PROTOCTL) Compatible to RPL_BOUNCE (RfC2812)

simpleAccess via the Properties Property of Type ServerProperties.

Includes all XML Documentation needed for the use of this feature. 

Note: This extension will be sent Upstream to the SmartIRC4Net Project.

git-svn-id: https://svn.origo.ethz.ch/huffelpuff-irc-bot/trunk@8 b439e43e-95c1-4580-a6c9-a09611c81d30